+++ src/translations/da.ts
@@ -8,6 +8,7 @@ const translation: Translation = {
 
   close: 'Luk',
   copy: 'Kopier',
+  currentValue: 'Nuværende regerer',
   progress: 'Status',
   remove: 'Fjerne',
   resize: 'Tipas størrelse',