+++ src/translations/he.ts
@@ -8,6 +8,7 @@ const translation: Translation = {
 
   close: 'סגור',
   copy: 'העתק',
+  currentValue: 'ערך נוכחי',
   progress: 'התקדמות',
   remove: 'לְהַסִיר',
   resize: 'שנה גודל',