+++ src/translations/ja.ts
@@ -8,6 +8,7 @@ const translation: Translation = {
 
   close: '閉じる',
   copy: 'コピー',
+  currentValue: '現在の価値',
   progress: '進行',
   remove: '削除',
   resize: 'サイズ変更',