+++ src/translations/nl.ts
@@ -8,6 +8,7 @@ const translation: Translation = {
 
   close: 'Sluiten',
   copy: 'Kopiëren',
+  currentValue: 'Huidige waarde',
   progress: 'Voortgang',
   remove: 'Verwijderen',
   resize: 'Formaat wijzigen',